# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é uma API de exemplo criada em Go com Gin e documentada com Swagger.",
        "title": "API de Exemplo com Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte da API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/articles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtém um artigo pelo seu ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Mostra um artigo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Artigo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Article"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Article": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Este é o conteúdo do artigo."
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Título do Artigo"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e o token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}